Moved popup functions to their own file
